refactor: integrate balance error handling in EarnTabs
