+++ src/components/icons/NetworkErrorIcon.tsx
@@ -1,0 +1,66 @@
+import type { SVGProps } from "react";
+
+interface NetworkErrorIconProps extends SVGProps<SVGSVGElement> {
+  color?: string;
+  size?: number;
+}
+
+export const NetworkErrorIcon = ({ color, size = 20, ...props }: NetworkErrorIconProps) => {
+  return (
+    <svg
+      version="1.0"
+      width={size}
+      height={size}
+      viewBox="0 0 512 512"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}>
+      <g transform="translate(0,512) scale(0.1,-0.1)" fill={color || "#FF1A1A"} stroke="none">
+        <path
+          d="M2450 5114 c-635 -40 -1197 -284 -1640 -711 -142 -137 -231 -240
+-335 -388 -221 -315 -359 -657 -426 -1049 -27 -155 -36 -484 -20 -652 73 -722
+451 -1382 1043 -1818 218 -160 504 -302 771 -382 81 -24 273 -68 343 -78 l41
+-6 17 128 c15 115 20 132 52 181 l36 53 -24 16 c-50 35 -154 141 -208 214
+-108 144 -201 327 -275 540 -19 55 -35 104 -35 109 0 5 100 9 230 9 l230 0 0
+160 0 160 -275 0 c-151 0 -275 1 -275 3 0 2 -11 68 -24 148 -28 162 -51 368
+-62 537 l-7 112 322 0 321 0 0 160 0 160 -321 0 -322 0 7 113 c11 169 34 376
+62 534 13 78 24 145 24 148 0 3 162 5 360 5 l360 0 0 -320 0 -320 155 0 155 0
+0 320 0 320 354 0 355 0 10 -52 c33 -165 71 -452 71 -540 l0 -48 48 0 c26 0
+97 3 157 6 l110 7 -3 41 c-8 113 -36 353 -53 451 -10 61 -19 116 -19 123 0 9
+91 12 420 12 l419 0 21 -47 c69 -156 128 -350 159 -518 l16 -90 153 19 153 20
+-7 45 c-12 94 -57 280 -95 398 -128 396 -341 736 -647 1035 -472 461 -1057
+710 -1717 731 -71 2 -155 3 -185 1z m-30 -809 l0 -465 -315 0 c-173 0 -315 1
+-315 3 0 2 13 41 29 88 126 369 325 673 522 798 35 23 67 41 72 41 4 0 7 -209
+7 -465z m403 415 c196 -131 360 -386 498 -772 l39 -108 -315 0 -315 0 0 465
+c0 275 4 465 9 465 5 0 43 -23 84 -50z m-1032 -129 c-110 -167 -241 -456 -302
+-664 -12 -44 -27 -81 -33 -83 -6 -2 -172 -3 -368 -2 l-358 3 74 96 c199 257
+517 515 793 641 101 46 226 97 240 98 7 0 -14 -40 -46 -89z m1649 37 c262
+-110 511 -278 720 -488 87 -87 218 -240 241 -282 9 -17 -11 -18 -354 -18
+l-364 0 -47 141 c-82 245 -168 430 -282 607 -30 45 -54 85 -54 88 0 8 34 -4
+140 -48z m-2077 -1240 c-25 -154 -49 -370 -59 -545 l-7 -123 -484 0 -483 0 4
+23 c3 12 8 51 11 87 16 173 91 451 165 618 l33 72 421 0 422 0 -23 -132z m-55
+-1169 c13 -182 40 -415 65 -548 l14 -71 -423 0 -422 0 -33 78 c-92 220 -145
+429 -172 680 l-5 42 482 0 482 0 12 -181z m176 -1006 c58 -207 206 -536 310
+-691 30 -45 53 -82 50 -82 -2 0 -45 16 -95 35 -196 76 -417 205 -594 348 -104
+84 -284 270 -363 375 l-63 82 368 0 368 0 19 -67z"
+        />
+        <path
+          d="M3722 2549 c-302 -22 -585 -157 -807 -383 -179 -183 -304 -430 -347
+-690 -18 -103 -14 -334 7 -431 67 -320 236 -592 483 -777 234 -177 479 -259
+772 -259 350 0 654 127 902 378 245 248 368 549 368 898 0 266 -82 522 -236
+735 -110 154 -304 320 -470 403 -190 96 -441 143 -672 126z m342 -333 c229
+-64 434 -206 555 -385 57 -84 116 -212 143 -308 32 -114 31 -372 0 -488 -48
+-174 -170 -371 -297 -479 -190 -161 -394 -236 -641 -236 -190 0 -343 43 -500
+139 -407 251 -567 754 -382 1199 118 283 379 498 688 567 88 20 351 14 434 -9z"
+        />
+        <path
+          d="M3427 1673 l-107 -108 142 -142 143 -143 -143 -143 -142 -142 110
+-110 110 -110 143 143 142 142 140 -140 c77 -77 144 -140 150 -140 6 0 57 47
+115 105 l105 105 -145 145 -145 145 145 145 145 145 -105 105 c-58 58 -109
+105 -115 105 -5 0 -73 -63 -150 -140 l-140 -140 -140 140 c-77 77 -142 140
+-145 140 -3 0 -53 -48 -113 -107z"
+        />
+      </g>
+    </svg>
+  );
+};
