+++ src/routes/popup/settings/wallets/[address]/qr.tsx
@@ -1,7 +1,7 @@
 import { useToasts, Section, useInput, Button, Input, Text } from "@arconnect/components-rebrand";
 import { useEffect, useState } from "react";
 import HeadV2 from "~components/popup/HeadV2";
-import { WarningIcon } from "~components/popup/Token";
+import { WarningIcon } from "~components/icons/WarningIcon";
 import browser from "webextension-polyfill";
 import { Degraded, WarningWrapper } from "~routes/popup/send";
 import { getKeyfile, type DecryptedWallet } from "~wallets";
