+++ src/routes/popup/subscriptions/subscriptionManagement.tsx
@@ -7,7 +7,7 @@ import styled from "styled-components";
 import { getSubscriptionData } from "~subscriptions";
 import { ButtonV2, ListItem } from "@arconnect/components";
 import { Degraded, WarningWrapper } from "../send";
-import { WarningIcon } from "~components/popup/Token";
+import { WarningIcon } from "~components/icons/WarningIcon";
 import Title from "~components/popup/Title";
 import type { CommonRouteProps } from "~wallets/router/router.types";
 import { useAsyncEffect } from "~utils/react/useAsyncEffect";
